% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process_multichain.R
\name{combine_multichain}
\alias{combine_multichain}
\title{combine_multichain}
\usage{
combine_multichain(
  datasets = list(),
  datasets_selected = c(1),
  burnin_values = NULL,
  end_values = NULL
)
}
\arguments{
\item{datasets}{List containing one or more datasets, generated by get_mcmc_datasets_multichain()}

\item{datasets_selected}{Vector of dataset numbers to be selected from list}

\item{burnin_values}{Vector of burnin values to use for selected datasets (set to 1 by default)}

\item{end_values}{Vector of end values to use for selected datasets (set to last value by default)}
}
\description{
Combine the post-burnin values from multiple Markov chains into one data frame
}
\details{
[TBA]
}
