% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create.R
\name{create_map}
\alias{create_map}
\title{create_map}
\usage{
create_map(
  shape_data = list(),
  param_values = c(),
  text_size = 1,
  display_axes = FALSE,
  border_colour_regions = "grey",
  ...
)
}
\arguments{
\item{shape_data}{Region shape data generated using map_shapes_load()}

\item{param_values}{Vector of parameter values for regions in shape_data}

\item{text_size}{Size of text to appear in legend and titles}

\item{display_axes}{TRUE/FALSE flag indicating whether to frame map and display latitude/longitude axes}

\item{border_colour_regions}{Colour to use for borders of regions. Set to NA if borders to be invisible.}

\item{...}{Additional optional parameters: \cr
   scale_manual Vector of scale intervals to use for param_values \cr
   colour_scale_manual Vector of colours with size greater than or equal to scale_manual - used to convert scale_manual to colours\cr
   lat_max, lat_min, long_max, long_min: borders if not to be set default \cr
   additional_border_shapes: Shape data for optional additional borders \cr
   border_colour_additional: colour to use for additional borders if any. \cr
'}
}
\description{
Create a map of one or more regions with colours denoting parameter values
}
\details{
Takes in region shape data generated using map_shapes_load() and parameter values for each region, plots map
  of the regions and fills regions with colour based on parameter values
}
