% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{data_match_single}
\alias{data_match_single}
\title{data_match_single}
\usage{
data_match_single(
  params = c(),
  input_data = list(),
  obs_sero_data = NULL,
  obs_case_data = NULL,
  ...
)
}
\arguments{
\item{params}{Values of input parameters in order FOI/FOI coefficients, R0/R0 coefficients, [etc.]}

\item{input_data}{List of population and vaccination data for multiple regions, with tables to cross-reference
with observed data, added using input_data_process}

\item{obs_sero_data}{Seroprevalence data for comparison, by region, year & age group, in format no. samples/no.
positives}

\item{obs_case_data}{Annual reported case/death data for comparison, by region and year, in format no. cases/no.
deaths}

\item{...}{= Additional parameters/flags/etc. (n_reps, mode_start, time_inc, enviro_data_const, enviro_data_var,
vaccine_efficacy, p_rep_severe, p_rep_death, m_FOI_Brazil, deterministic, mode_time,
mode_parallel, cluster, p_severe_inf, p_death_severe_inf)}
}
\description{
Function which runs the model to create simulated data corresponding to supplied observed data for a
  single set of parameters with one or more repetitions
}
\details{
This function takes in observed data and produces a corresponding simulated dataset for comparison using
  supplied model parameters (in the params input variable), population and vaccination settings (in the
  input_data input variable) and other settings including parameter type, time increment and environmental covariate
  values (in the const_list input variable). The simulated dataset will be able to be compared directly to the
  observed dataset - individual data values will be produced for the same years and regions (or combinations of
  regions), seroprevalence data will be produced for the same age groups, etc.
}
