% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process_multichain.R
\name{get_mcmc_additional_params_multichain}
\alias{get_mcmc_additional_params_multichain}
\title{get_mcmc_additional_params_multichain}
\usage{
get_mcmc_additional_params_multichain(
  datasets = c(),
  datasets_selected = c(1),
  burnin_values = NULL,
  end_values = NULL,
  parameters = c("vaccine_efficacy")
)
}
\arguments{
\item{datasets}{List containing one or more datasets, generated by get_mcmc_datasets_multichain()}

\item{datasets_selected}{Vector of dataset numbers to be selected from list}

\item{burnin_values}{Vector of burn-in iteration values (1 per dataset) before which data discarded (defaults to 1)}

\item{end_values}{Vector of end iteration values (1 per dataset) after which data discarded (defaults to chain ends)}

\item{parameters}{Vector of names of parameters to be extracted}
}
\description{
Get case reporting and vaccine efficacy data from MCMC results
}
\details{
This function extracts values (where included in calculations) of the probabilities of severe and fatal
  cases being reported and the vaccine efficacy from datasets generated by get_mcmc_datasets_multichain(),
  returning a list of datasets.
}
