% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{get_mcmc_FOI_R0_data}
\alias{get_mcmc_FOI_R0_data}
\title{get_mcmc_FOI_R0_data}
\usage{
get_mcmc_FOI_R0_data(
  input_frame = list(),
  enviro_data_const = list(),
  enviro_data_var = NULL
)
}
\arguments{
\item{input_frame}{Data frame of MCMC output data produced by get_mcmc_data(), truncate_mcmc_data() and/or
combine_multichain() OR data frame containing parameter values as columns with parameter names as column headings}

\item{enviro_data_const}{TBA}

\item{enviro_data_var}{TBA
'}
}
\description{
Extract spillover force of infection (FOI) and reproduction number (R0) values from MCMC output data
}
\details{
This function takes in a data frame produced by functions get_mcmc_data(), truncate_mcmc_data() and/or
combine_multichain() and calculates spillover force of infection (FOI) and reproduction number (R0) values for
each row in the frame (representing points on the chain) based on the type of fit carried out.

TBA
}
