% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_dist.R
\name{mcmc_datasets_sort_by_burden}
\alias{mcmc_datasets_sort_by_burden}
\title{mcmc_datasets_sort_by_burden}
\usage{
mcmc_datasets_sort_by_burden(input_frame = list())
}
\arguments{
\item{input_frame}{Data frame of MCMC output data produced by get_mcmc_data(), truncate_mcmc_data() and/or
  combine_multichain()
'}
}
\description{
[TBA]
}
\details{
[TBA]
}
