% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_dist.R
\name{mcmc_R0_value_probs}
\alias{mcmc_R0_value_probs}
\title{mcmc_R0_value_probs}
\usage{
mcmc_R0_value_probs(FOI_R0_values = list(), R0_target_values = c(1))
}
\arguments{
\item{FOI_R0_values}{Data frame of spillover FOI and/or R0 values obtained for multiple
regions using get_mcmc_FOI_R0_data()}

\item{R0_target_values}{Vector of target values of R0
'}
}
\description{
Get sets of values by region of probability that R0 equals or exceeds one or more values
}
\details{
[TBA]
}
