% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_region_input_data}
\alias{plot_region_input_data}
\title{plot_region_input_data}
\usage{
plot_region_input_data(
  input_data = list(),
  region = "",
  data_type = "vacc",
  colour_scale = c(),
  output_filename = NULL,
  pixel_width = 480,
  pixel_height = 480
)
}
\arguments{
\item{input_data}{Input data list in standard format}

\item{region}{Selected region from input_data$region_labels}

\item{data_type}{Type of data to plot, either "vacc" or "pop"}

\item{colour_scale}{TBA}

\item{output_filename}{TBA}

\item{pixel_width}{TBA}

\item{pixel_height}{TBA
'}
}
\description{
TBA
}
\details{
TBA
}
