% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match2.R
\name{data_match_multi2}
\alias{data_match_multi2}
\title{data_match_multi2}
\usage{
data_match_multi2(
  param_sets = list(),
  input_data = list(),
  env_covar_values = list(),
  template = list(),
  fixed_extra = list(),
  ...
)
}
\arguments{
\item{param_sets}{Data frame of values of proposed parameters, one set per row, with parameter names as headings}

\item{input_data}{List of population and vaccination data for multiple regions, with tables to cross-reference
with observed data, added using input_data_process2}

\item{env_covar_values}{TBA}

\item{template}{TBA}

\item{fixed_extra}{TBA}

\item{...}{= Constant additional parameters/flags/etc. (n_reps, mode_start, time_inc, enviro_data_const, enviro_data_var,
vaccine_efficacy, p_rep_severe, p_rep_death, m_FOI_BRA, deterministic, mode_time,
mode_parallel, cluster, p_severe_inf, p_death_severe_inf)}
}
\description{
Function which runs the model to create simulated data corresponding to supplied observed data for
  multiple parameter sets
}
\details{
This function runs the data_match_single2() function for multiple parameter sets. It takes in observed data
  and produces corresponding simulated datasets for comparison using multiple sets of supplied model parameters (in
  the param_sets input variable), population and vaccination settings (in the input_data input variable) and other
  settings including parameter type, time increment and environmental covariate values (in the consts input
  variable). The simulated dataset will be able to be compared directly to the observed dataset - individual data
  values will be produced for the same years and regions (or combinations of regions), seroprevalence data will be
  produced for the same age groups, etc.
}
