% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_dist.R
\name{get_FOI_R0_dist_data}
\alias{get_FOI_R0_dist_data}
\title{get_FOI_R0_dist_data}
\usage{
get_FOI_R0_dist_data(FOI_R0_values = list())
}
\arguments{
\item{FOI_R0_values}{Data frame of spillover FOI and/or R0 values obtained for multiple
  regions using get_mcmc_FOI_R0_data()
'}
}
\description{
[TBA]
}
\details{
[TBA]
}
