% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_results_process.R
\name{get_mcmc_enviro_coeff_data}
\alias{get_mcmc_enviro_coeff_data}
\title{get_mcmc_enviro_coeff_data}
\usage{
get_mcmc_enviro_coeff_data(input_frame = list(), type = "FOI+R0 enviro")
}
\arguments{
\item{input_frame}{Data frame of MCMC output data produced by get_mcmc_data(), truncate_mcmc_data() and/or
combine_multichain() OR data frame containing parameter values as columns with parameter names as column headings}

\item{type}{Type of parameter set (FOI and/or R0 coefficients associated with environmental
  covariates); choose from "FOI enviro", "FOI+R0 enviro"
'}
}
\description{
Extract values of environmental coefficients from MCMC output data
}
\details{
This function takes in a data frame produced by functions get_mcmc_data() and/or truncate_mcmc_data() and
calculates values of the coefficients of environmental covariates for each row in the frame
(representing points on the chain) based on the type of fit carried out.
}
